+++ ComtainerDrawingApi/ContainerDrawingApi.v2/Models/Container/Container.cs
@@ -13,11 +13,11 @@ namespace ContainerDrawingApi.v2.Models
 
         public string name { get; set; }
 
-        public int length { get; set; }
+        public double length { get; set; }
 
-        public int width { get; set; }
+        public double width { get; set; }
 
-        public int height { get; set; }
+        public double height { get; set; }
 
         public double maxAllowedVolume { get; set; }
